added exception handling for not found ids and blank to-do inputs

+++ src/main/java/com/cedrus/practice/todo/Spring_Boot_Todos/service/ToDoService.java
@@ -1,5 +1,7 @@
 package com.cedrus.practice.todo.Spring_Boot_Todos.service;
 
+import com.cedrus.practice.todo.Spring_Boot_Todos.exception.BlankToDoException;
+import com.cedrus.practice.todo.Spring_Boot_Todos.exception.ToDoNotFoundException;
 import com.cedrus.practice.todo.Spring_Boot_Todos.model.ToDoItem;
 import com.cedrus.practice.todo.Spring_Boot_Todos.repository.ToDoRepo;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,23 +19,33 @@ public class ToDoService {
         return toDoRepo.findAll();
     }
 
-    public void addToDo(ToDoItem toDoItem) {
+    public void addToDo(ToDoItem toDoItem) throws BlankToDoException {
+        if (toDoItem.getItem().length() < 1)
+            throw new BlankToDoException("No text entered for to do!");
+
         toDoRepo.save(toDoItem);
         System.out.println("ToDo item added to DB");
     }
 
-    public void updateToDo(int id, ToDoItem newToDo) {
+    public void updateToDo(int id, ToDoItem newToDo) throws ToDoNotFoundException, BlankToDoException {
         ToDoItem toBeUpdated = toDoRepo.findById(id);
 
-        if (newToDo.getItem().length() > 0) {
-            toBeUpdated.setItem(newToDo.getItem());
-            toDoRepo.save(toBeUpdated);
-        }
-        else
-            System.out.println("no updated message entered");
+        if (toBeUpdated == null)
+            throw new ToDoNotFoundException(String.format("ToDo not found for id: " + id));
+
+        if (newToDo.getItem().length() < 1)
+            throw new BlankToDoException("No text entered for to do");
+
+        toBeUpdated.setItem(newToDo.getItem());
+        toDoRepo.save(toBeUpdated);
     }
 
-    public void removeToDo(int id) {
+    public void removeToDo(int id) throws ToDoNotFoundException {
+        ToDoItem toBeDeleted = toDoRepo.findById(id);
+
+        if (toBeDeleted == null)
+            throw new ToDoNotFoundException(String.format("Todo not found for id: " + id));
+
         toDoRepo.deleteById(id);
     }
 }
